+++ CHANGES.md
@@ -1,12 +1,15 @@
 ## CHANGES
 
 ### next
+  * ...
+
+### 0.8.20260323
   * NEW check: simple IPv6 Path MTU checks against RIPE Atlas anchors
   * CHANGED: move root DNS server config to the ini file
   * CHANGED: support "ping is expected to fail" per root server and AF
   * CHANGED: all defined open DNS resolvers are now in one check
   * NEW: added "select all checks" to GUI
-	* CHANGED: SSH hos key checks can now have multiple keys per host, any mathces are accepted
+	* CHANGED: SSH host key checks can now have multiple keys per host, any match is accepted
 
 ### 0.7.20260311
   * NEW check: SSH host key checks